build: abstracting functions
